fix theo comment cua trainer
