+++ prisma/schema.prisma
@@ -16,7 +16,7 @@ datasource db {
 model User {
   id        Int   @id @default(autoincrement())
   email     String   @unique
-  name      String?
+  name      String
   password  String
   bio       String?
   image     String?
