+++ src/auth/dto/register.dto.ts
@@ -12,12 +12,3 @@ export class RegisterDto {
   bio?: string;
   image?: string;
 }
-
-export class LoginDto {
-  @IsEmail()
-  email: string;
-
-  @IsNotEmpty()
-  @MinLength(6)
-  password: string;
-}
